Partial revert of "refactoring to support coreos-cluster configuration. Moved ssh_config generation and basic python env installation from ansible-deployment to ansible-coreos-cluster".
We need bootstrap in deployment too, in case there are auto-scale groups and new node is selected as cluster manager. In that case it will not have basic python env. Bootstrap only the coreos_cluster_manager, not whole coreos_cluster as previously.

ssh_config is reused from ansible-coreos-cluster. If auto-scale group results in more nodes, it may need to be regenerated.